Create macOSUpgrades_CheckCachedInstaller
A script to check to see if a cached macOS installer exists, verify it has has downloaded completely, or re download if either of the previous 2 conditions are not met.
Designed to be used BEFORE running the macOSUpgrades or macOSUpgrades_SelfService scripts. Once those scripts are enabled this script should be stopped.